* amd/nfs_start.c (checkup): cast getpagesize() return val to long, to avoid conflicts in division of a long by an int (on hpux11i, with their ANSI/C compiler).
* minor new ports: ia64-hp-hpux11.20, i386-unknown-freebsd5.0,
i386-unknown-freebsd4.4, i386-pc-linux-rh7.2,
ia64-unknown-linux-rh7.1.

* include/am_defs.h: actually declare extern definition for
xdr_callmsg() if it doesn't already exist.

* aux/macros/header_templates.m4: declare template for
HAVE_EXTERN_XDR_CALLMSG.

* configure.in: check for extern definition for xdr_callmsg()
because hpux11 has that function but a bad <rpc/xdr.h> header file
that doesn't define it in the right place.

* libamu/mount_fs.c (print_nfs_args): cast ->rdev to u_long b/c
on hpux11 its type is dev_t.

* conf/nfs_prot/nfs_prot_hpux11.h: don't support autofs on hpux11
yet.

+++ amd/nfs_start.c
@@ -38,7 +38,7 @@
  *
  *      %W% (Berkeley) %G%
  *
- * $Id: nfs_start.c,v 1.8 2001/01/10 03:22:15 ezk Exp $
+ * $Id: nfs_start.c,v 1.9 2001/11/27 03:11:25 ezk Exp $
  *
  */
 
@@ -83,7 +83,7 @@ checkup(void)
 #ifdef HAVE_GETPAGESIZE
     dlog("%#lx bytes of memory allocated; total is %#lx (%ld pages)",
 	 (long) (next_mem - max_mem), (unsigned long) next_mem,
-	 ((long) next_mem + getpagesize() - 1) / getpagesize());
+	 ((long) next_mem + getpagesize() - 1) / (long) getpagesize());
 #else /* not HAVE_GETPAGESIZE */
     dlog("%#lx bytes of memory allocated; total is %#lx",
 	 (long) (next_mem - max_mem), (unsigned long) next_mem);
